Refactor PatchTST model structure and add registry
