+++ LGHackerton/models/__init__.py
@@ -1,2 +1,2 @@
 """Model training components."""
-from .model_registry import ModelRegistry
+from .registry import ModelRegistry
